test : 토큰 테스트 수정
토큰 테스트 리팩터링 클래스명, 테스트 분리
