+++ waldreg/service-layer/auth/src/test/java/org/waldreg/token/publisher/SuperPublisherTest.java
@@ -1,0 +1,36 @@
+package org.waldreg.token.publisher;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.waldreg.token.dto.TokenDto;
+import org.waldreg.token.jwt.publisher.JwtTokenPublisher;
+
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes = {SuperTokenPublisher.class})
+public class SuperPublisherTest{
+
+    @Autowired
+    private TokenPublisher superTokenPublisher;
+
+
+    @Test
+    @DisplayName("super 토큰 생성 성공")
+    public void CREATE_SUPER_TOKEN_SUCCESS_TEST(){
+        //given
+        int id = 1;
+        TokenDto tokenDto = TokenDto.builder().id(id).build();
+
+        //when
+
+        //then
+        Assertions.assertDoesNotThrow(superTokenPulisher.publish(tokenDto));
+
+    }
+
+
+}
